// src/generated/contracts/navigation/NavigationTarget.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NavigationDestination } from "./NavigationDestination";
import type { NavigationSelection } from "./NavigationSelection";

/**
 * A versioned destination with an optional selected object.
 */
export type NavigationTarget = { version: 1, destination: NavigationDestination, selection?: NavigationSelection, };